fix: Correct empty datalist on initial load
When loading the site for the first time, the app fetches bart station information from the bart api, saves it to indexeddb, and binds that data to the datalist. The problem is on that initial load, the data wasn't being read from indexeddb. My fix was to bind the response from the bart api to the datalist for the initial load; any loads after that would request data from indexeddb. Datalist now works as expected.