bug: use proxy server for bart api
The BART api is accessed through an HTTP connection, and this app is served through HTTPS; this led to Cross-Origin Resource Sharing errors when making any BART ajax request. Typically, fixing this requires reconfiguring the API's server for CORS, however that option is unavailable to me. My best alternative proved to be the use of a proxy server. Instead of my BART requests going directly to the BART API, those requests are sent to the proxy - which is configured for CORS - which then makes the request to BART, and sends the reponse back to my app.